Add files via upload
The objective of this analysis was to perform an exploratory data analysis (EDA) on a Spotify dataset to understand the distribution and relationships of various features related to songs and user preferences. The dataset included various attributes such as track_name, artist, popularity, danceability, energy, speechiness, and a target variable indicating user preference.